+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.9.3] - 2022-10-03
+### Fixed
+- Fixed a bug in showing team-activity-alert
+- If there is an error in getting team-activity, trying again after 5 minutes
+
 ## [3.9.2] - 2022-10-01
 ### Added
 - Logging YAML error while reading file content